Branch3 finishing product and category product

+++ backend/src/controllers/category_controller.js
@@ -7,9 +7,14 @@ const fs = require("fs")
 const { categorySchema } = require("../helper/validation_schema")
 
 module.exports.getCategory = async (req, res) => {
-  try {
 
-    const GET_CATEGORY = `SELECT * FROM categories;`
+  const per_page = Number(req.query.per_page) || 3
+  const current_page = Number(req.query.current_page) || 1
+  const search = req.query.search || ""
+  try {
+    const offset = (current_page - 1) * per_page
+
+    const GET_CATEGORY = `SELECT * FROM categories WHERE categoryName LIKE '%${search}%' LIMIT ${offset}, ${per_page};`
 
     const [categories] = await db.execute(GET_CATEGORY)
 
@@ -23,73 +28,13 @@ module.exports.getCategory = async (req, res) => {
       throw new createError(httpStatus.BAD_REQUEST, 'kategori tidak ditemukan')
     }
 
-    const responseStatus = new createResponse(
-      httpStatus.OK,
-      'success', true, 1, 1, categories
-    )
+    const TOTAL_CATEGORY = `SELECT COUNT (*) as total_category FROM categories WHERE categoryName LIKE '%${search}%';`
 
-    res.status(responseStatus.status).send(responseStatus)
-
-  } catch (error) {
-    console.log(error)
-  }
-}
-
-module.exports.getCategoryByName = async (req, res) => {
-
-  const categoryName = req.params.name
-
-  try {
-
-    const GET_CATEGORY_BY_NAME = `SELECT * FROM categories WHERE categoryName LIKE '%${categoryName}%';`
-
-    const [categories] = await db.execute(GET_CATEGORY_BY_NAME)
-
-    if (!categories.length) {
-      const responseStatus = new createResponse(
-        httpStatus.BAD_REQUEST,
-        'Error', false, 1, 1, 'Kategori tidak ditemukan'
-      )
-
-      res.status(responseStatus.status).send(responseStatus)
-      throw new createError(httpStatus.BAD_REQUEST, 'kategori tidak ditemukan')
-    }
+    const [totalCategory] = await db.execute(TOTAL_CATEGORY)
 
     const responseStatus = new createResponse(
       httpStatus.OK,
-      'success', true, 1, 1, categories
-    )
-
-    res.status(responseStatus.status).send(responseStatus)
-
-  } catch (error) {
-    console.log(error)
-  }
-}
-
-module.exports.getCategoryById = async (req, res) => {
-
-  const idCategory = req.params.id
-
-  try {
-
-    const GET_CATEGORY_BY_ID = `SELECT * FROM categories WHERE id = ${idCategory};`
-
-    const [categories] = await db.execute(GET_CATEGORY_BY_ID)
-
-    if (!categories.length) {
-      const responseStatus = new createResponse(
-        httpStatus.BAD_REQUEST,
-        'Error', false, 1, 1, 'Kategori tidak ditemukan'
-      )
-
-      res.status(responseStatus.status).send(responseStatus)
-      throw new createError(httpStatus.BAD_REQUEST, 'kategori tidak ditemukan')
-    }
-
-    const responseStatus = new createResponse(
-      httpStatus.OK,
-      'success', true, 1, 1, categories
+      'success', true, totalCategory[0], 1, categories
     )
 
     res.status(responseStatus.status).send(responseStatus)
@@ -209,6 +154,19 @@ module.exports.deleteCategory = async (req, res) => {
 
       res.status(responseStatus.status).send(responseStatus)
       throw new createError(httpStatus.BAD_REQUEST, 'kategori tidak ditemukan')
+    }
+
+    const CHECK_USAGE_CATEGORY = `SELECT * FROM products WHERE idCategory = ?; `
+    const [isUsage] = await db.execute(CHECK_CATEGORY, [idCategory])
+
+    if (isUsage.length) {
+      const responseStatus = new createResponse(
+        httpStatus.BAD_REQUEST,
+        'Error', false, 1, 1, 'Category is used by some products'
+      )
+
+      res.status(responseStatus.status).send(responseStatus)
+      throw new createError(httpStatus.BAD_REQUEST, 'Category is used by some products')
     }
 
     const DELETE_CATEGORY = `DELETE FROM categories WHERE id = ?;`
